[touch:10586]{t:10}dont typecast result of filter_var to int just yet, it could have produced an empty string which is significantly different from int 0

+++ core/libraries/form_sections/strategies/normalization/EE_Int_Normalization.strategy.php
@@ -45,7 +45,7 @@ class EE_Int_Normalization extends EE_Normalization_Strategy_Base
                 )
             );
         }
-        $value_to_normalize = (int)filter_var(
+        $value_to_normalize = filter_var(
             $value_to_normalize,
             FILTER_SANITIZE_NUMBER_FLOAT,
             FILTER_FLAG_ALLOW_FRACTION
